Cleanup encode and decode Cleanup index Fix abiEncode, it uses abiCoder from ethers now!
