+++ package.json
@@ -24,7 +24,6 @@
     "eslint": "^8.23.1",
     "eslint-config-google": "^0.14.0",
     "ethers": "^5.7.0",
-    "solc-typed-ast": "^10.1.6",
     "starknet": "^4.4.2",
     "hardhat": "^2.0.0",
     "mocha": "^7.1.2",
