+++ src/ethers/Contract.ts
@@ -29,19 +29,16 @@ import {
   EventFilter,
   Listener,
   Provider,
-  Log,
   TransactionRequest,
   TransactionResponse,
   Block,
   TransactionReceipt
 } from "@ethersproject/abstract-provider";
 import {id as keccak} from "@ethersproject/hash";
-import { parse, TypeNode } from "solc-typed-ast";
-import { decode, decodeEvents, decode_, encodeValueOuter, SolValue } from "../encode";
+import { abiCoder, decode, decodeEvents, decode_, encode } from "../transcode";
 import { FIELD_PRIME } from "starknet/dist/constants";
 import {readFileSync} from "fs";
 import {normalizeAddress} from "../utils";
-import {abiEncode} from "../abiEncode";
 
 const ASSERT_ERROR = "An ASSERT_EQ instruction failed";
 
@@ -221,21 +218,6 @@ export class WarpContract extends EthersContract {
     throw new Error("Not implemented yet");
   }
 
-  public argStringifier(arg: any): SolValue {
-    return Array.isArray(arg) ? arg.map(this.argStringifier) : arg.toString();
-  }
-
-  private format(paramType: ParamType): string {
-    if (paramType.type === "tuple") {
-      return `tuple(${paramType.components.map(this.format).join(",")})`;
-    } else if (paramType.arrayChildren !== null) {
-      return `${this.format(paramType.arrayChildren)}[${
-        paramType.arrayLength >= 0 ? paramType.arrayLength : ""
-      }]`;
-    } else {
-      return paramType.type;
-    }
-  }
   private buildDefault(solName: string, fragment: FunctionFragment) {
     if (fragment.constant) {
       return this.buildCall(solName, fragment);
@@ -244,25 +226,15 @@ export class WarpContract extends EthersContract {
   }
 
   private buildInvoke(solName: string, fragment: FunctionFragment) {
-    const inputTypeNodes = fragment.inputs.map((tp) => {
-      const res = parse(this.format(tp), {
-        ctx: undefined,
-        version: undefined,
-      }) as TypeNode;
-      return res;
-    });
     const cairoFuncName =
       solName + "_" + this.interface.getSighash(fragment).slice(2); // Todo finish this keccak (use web3)
 
     return async (...args: any[]) => {
       console.log({cairoFuncName})
-      const calldata = args.flatMap((arg, i) =>
-        encodeValueOuter(
-          inputTypeNodes[i],
-          this.argStringifier(arg),
-          "we don't care"
-        )
-      );
+      const calldata = encode(
+        fragment.inputs,
+        args,
+      )
       try {
         console.log("INVOKE FUNCTION")
         const invokeResponse = await this.starknetContract.providerOrAccount.invokeFunction(
@@ -272,11 +244,12 @@ export class WarpContract extends EthersContract {
             entrypoint: cairoFuncName,
           },
           {
-            maxFee: FIELD_PRIME,
+            // Set maxFee to some high number for goerli
+            maxFee: process.env.STARKNET_PROVIDER_BASE_URL ? undefined : (2n ** 250n).toString()
           }
         );
         console.log("Before to etheresTransaction")
-        const abiEncodedInputs = abiEncode(fragment.inputs, args.map(a => this.argStringifier(a)))
+        const abiEncodedInputs = abiCoder.encode(fragment.inputs, args.map(a => this.argStringifier(a)))
         const sigHash = this.ethersContractFactory.interface.getSighash(fragment);
         const data = sigHash.concat(abiEncodedInputs.substring(2));
         return this.toEtheresTransactionResponse(
@@ -296,24 +269,13 @@ export class WarpContract extends EthersContract {
     };
   }
   private buildCall(solName: string, fragment: FunctionFragment) {
-    const inputTypeNodes = fragment.inputs.map(
-      (tp) =>
-        parse(this.format(tp), {
-          ctx: undefined,
-          version: undefined,
-        }) as TypeNode
-    );
-
     const cairoFuncName =
       solName + "_" + this.interface.getSighash(fragment).slice(2); // Todo finish this keccak (use web3)
     // @ts-ignore
     return async (...args: any[]) => {
-      const calldata = args.flatMap((arg, i) =>
-        encodeValueOuter(
-          inputTypeNodes[i],
-          this.argStringifier(arg),
-          "we don't care"
-        )
+      const calldata = encode(
+        fragment.inputs,
+        args,
       );
       try {
         const output_before = await this.starknetContract.providerOrAccount.callContract(
@@ -349,8 +311,10 @@ export class WarpContract extends EthersContract {
     // @ts-ignore
     this[solName] = this.buildDefault(solName, fragment);
 
+    // TODO: functions have a slightly different return type for single value returns
     this.functions[solName] = this.buildDefault(solName, fragment);
 
+    // TODO: callStatic have a slightly different return type for single value returns
     this.callStatic[solName] = this.buildCall(solName, fragment);
   }
 
@@ -472,7 +436,7 @@ export class WarpContract extends EthersContract {
           removed: false,
           address: normalizeAddress(e.from_address),
           // abi encoded data
-          data: abiEncode(eventFragment.inputs, resultsArray as SolValue[]),
+          data: abiCoder.encode(eventFragment.inputs, resultsArray),
           topics: [selector],
           transactionHash,
           logIndex: i,
